requirements.txt
google-adk==1.2.1
cloud-sql-python-connector[pg8000]